rag/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.0.0
plotly>=5.15.0
